requirements.txt
Django==1.11.1
jsonfield
django_admin_bootstrapped
django-geoip
django-tastypie
dj_database_url
#django-progressbarupload # has issues with simple json
psycopg2==2.7
pykml
gunicorn
pyelasticsearch
django-haystack
#haystack  # apparently the order matters... maybe: elastic, haystack, dj-haystack?
BeautifulSoup
html
gpxpy
shortuuid
python-geoip
python-geoip-geolite2
oauth2client